Added vimrc to list of files and check for whether or not a given file has a symlink

+++ setup_dotfiles.sh
@@ -12,32 +12,36 @@
 # dotfiles directory
 DOTFILE_DIR=~/Project/dotfiles
 # dotfiles backup directory
-BACKUP_DIR=$dir/dotfiles_backup
+BACKUP_DIR=$DOTFILE_DIR/dotfiles_backup
 # list of files/folders to symlink in the home directory (~/)
-DOTFILES="bashrc zshrc gitignore_global"
+DOTFILES="vimrc bashrc zshrc gitconfig gitignore_global"
 
 echo "Running dotfiles setup..."
 
 # Create backup directory
 echo -n "Creating $BACKUP_DIR directory for backup of any existing dotfiles in the home directory"
-mkdir -p $BACKUP_DIR
+mkdir -p ~/Project/dotfiles/dotfiles_backup
 echo "...SUCCESS"
 
-# Move any existing dotfiles in the given $DOTFILES list from the home directory to the backup directory,
-# and create a symlink from the home directory to any of the dotfiles in the dotfiles directory
+# Move any existing dotfiles that is not a symlink in the given $DOTFILES list from
+# the home directory to the backup directory, and create a symlink from the home directory
+# to any of the dotfiles in the dotfiles directory
 for file in $DOTFILES
 do
-	if [ -f ~/$file ]
+	if [ -f ~/.$file ] && [ ! -L ~/.$file ]
 	then
 		echo -n "Moving .$file from ~ to $BACKUP_DIR"
 		mv ~/.$file $BACKUP_DIR
 		echo "...SUCCESS"
 	fi
 	
-	echo -n "Creating symlink to .$file in home directory
-	ln -s $DOTFILE_DIR/.$file ~/.$file
-	echo "...SUCCESS"
+	if [ ! -L ~/.$file ] && [ -f $DOTFILE_DIR/.$file ]
+	then
+		echo -n "Creating symlink to .$file in home directory"
+		ln -s $DOTFILE_DIR/.$file ~/.$file
+		echo "...SUCCESS"
+	fi
 done
 
-echo "dotfiles setup complete!~"
-exit 0 
+echo "dotfiles setup complete!"
+exit 0
